+++ simpletransformers/ner/ner_model.py
@@ -56,15 +56,7 @@ import wandb
 
 
 class NERModel:
-    def __init__(
-        self,
-        model_type,
-        model_name,
-        labels=None,
-        args=None,
-        use_cuda=True,
-        cuda_device=-1,
-    ):
+    def __init__(self, model_type, model_name, labels=None, args=None, use_cuda=True, cuda_device=-1, **kwargs):
         """
         Initializes a NERModel
 
@@ -115,7 +107,7 @@ class NERModel:
 
         config_class, model_class, tokenizer_class = MODEL_CLASSES[model_type]
 
-        self.model = model_class.from_pretrained(model_name, num_labels=self.num_labels)
+        self.model = model_class.from_pretrained(model_name, num_labels=self.num_labels, **kwargs)
 
         if use_cuda:
             if torch.cuda.is_available():
@@ -143,9 +135,9 @@ class NERModel:
         if args:
             self.args.update(args)
 
-        self.tokenizer = tokenizer_class.from_pretrained(
-            model_name, do_lower_case=self.args["do_lower_case"]
-        )
+
+        self.tokenizer = tokenizer_class.from_pretrained(model_name, do_lower_case=self.args['do_lower_case'], **kwargs)
+
 
         self.args["model_name"] = model_name
         self.args["model_type"] = model_type
@@ -474,10 +466,8 @@ class NERModel:
             ) and not os.path.exists(output_dir_current):
                 os.makedirs(output_dir_current)
 
-            if (
-                args["save_model_every_epoch"]
-                and epoch_number != args["num_train_epochs"]
-            ):
+            if args['save_model_every_epoch']:
+
                 model_to_save = model.module if hasattr(model, "module") else model
                 model_to_save.save_pretrained(output_dir_current)
                 self.tokenizer.save_pretrained(output_dir_current)