+++ simpletransformers/question_answering/question_answering_model.py
@@ -62,9 +62,9 @@ import wandb
 
 
 class QuestionAnsweringModel:
-    def __init__(
-        self, model_type, model_name, args=None, use_cuda=True, cuda_device=-1
-    ):
+
+    def __init__(self, model_type, model_name, args=None, use_cuda=True, cuda_device=-1, **kwargs):
+
         """
         Initializes a QuestionAnsweringModel model.
 
@@ -89,7 +89,7 @@ class QuestionAnsweringModel:
         }
 
         config_class, model_class, tokenizer_class = MODEL_CLASSES[model_type]
-        self.model = model_class.from_pretrained(model_name)
+        self.model = model_class.from_pretrained(model_name, **kwargs)
 
         if use_cuda:
             if torch.cuda.is_available():
@@ -125,9 +125,9 @@ class QuestionAnsweringModel:
         if args:
             self.args.update(args)
 
-        self.tokenizer = tokenizer_class.from_pretrained(
-            model_name, do_lower_case=self.args["do_lower_case"]
-        )
+
+        self.tokenizer = tokenizer_class.from_pretrained(model_name, do_lower_case=self.args['do_lower_case'], **kwargs)
+
 
         self.args["model_name"] = model_name
         self.args["model_type"] = model_type
@@ -542,10 +542,8 @@ class QuestionAnsweringModel:
             ) and not os.path.exists(output_dir_current):
                 os.makedirs(output_dir_current)
 
-            if (
-                args["save_model_every_epoch"]
-                and epoch_number != args["num_train_epochs"]
-            ):
+            if args['save_model_every_epoch']:
+
                 model_to_save = model.module if hasattr(model, "module") else model
                 model_to_save.save_pretrained(output_dir_current)
                 self.tokenizer.save_pretrained(output_dir_current)